Work for floats as well

+++ calc.py
@@ -1,8 +1,9 @@
 # Perform simple arithmetic encoded in an input string:
 # '1 + 2' -> 3, or '1 - 2' -> -1.
+# Hello world
 def compute(expression):
     num0, operator, num1 = expression.split(' ')
-    num0, num1 = int(num0), int(num1)
+    num0, num1 = float(num0), float(num1)
     if operator == '+':
         return num0 + num1
     elif operator == '-':
